Pull all the logic into support
Leave default final imple empty. This gives implementors
chance to redefine default behavior more easily.
